The share of learners in each course based on specific characteristics.

The share of learners in each course based on specific characteristics.

crsra_membershares(
  all_tables,
  groupby = c("roles", "country", "language", "gender", "empstatus", "education",
    "stustatus"),
  remove_missing = TRUE
)

Arguments

all_tables
A list from crsra_import_course or crsra_import

groupby
A character string indicating the how to break down learners in each course. The default is set to roles and returns the share of students in each category such as Learner, Not Enrolled, Pre-Enrolled Learner, Mentor, Browser, and Instructor. Other values are country (for grouping based on country), language (for grouping based on language), gender (for grouping by gender), education (for grouping by education level), stustatus (for grouping by student status), empstatus (for grouping by employment status), and country (for grouping by country). Note that this grouping uses the entries in the table users that is not fully populated so by grouping you lose some observations.

remove_missing
Should the NA be removed from the groupby column?

Value

A table which indicates the total number and the share of students in each group for each course

Examples

crsra_membershares(
example_course_import,
groupby = "country")

#> $`Fake Course`
#> # A tibble: 69 x 3
#>    y     Total Share
#>    <chr> <int> <dbl>
#>  1 US      143  0.31
#>  2 IN       53  0.12
#>  3 GB       19  0.04
#>  4 CN       17  0.04
#>  5 AU       15  0.03
#>  6 BR       15  0.03
#>  7 CA       13  0.03
#>  8 DE       11  0.02
#>  9 FR       11  0.02
#> 10 ES       10  0.02
#> # … with 59 more rows
#> 

crsra_membershares(
example_course_import,
groupby = "roles", remove_missing = FALSE)

#> $`Fake Course`
#> # A tibble: 4 x 3
#>   y                    Total Share
#>   <chr>                <int> <dbl>
#> 1 NA                     453  0.98
#> 2 LEARNER                  9  0.02
#> 3 BROWSER                  1  0   
#> 4 PRE_ENROLLED_LEARNER     1  0   
#> 

crsra_membershares(
example_course_import,
groupby = "roles", remove_missing = TRUE)

#> $`Fake Course`
#> # A tibble: 3 x 3
#>   y                    Total Share
#>   <chr>                <int> <dbl>
#> 1 LEARNER                  9  0.82
#> 2 BROWSER                  1  0.09
#> 3 PRE_ENROLLED_LEARNER     1  0.09
#>