Returns a list of tables a variable appears in

Returns a list of tables a variable appears in

crsra_whichtable(all_tables, col_name)

Arguments

all_tables
A list from crsra_import_course or crsra_import

col_name
The name of the column/variable to look for

Value

A list of tables that a specific variable appears in

Examples

crsra_whichtable(example_course_import, "assessment_id")

#> $`Fake Course`
#> [1] "assessment_actions"               "assessment_assessments_questions"
#> [3] "assessment_responses"             "assessments"                     
#> [5] "course_branch_item_assessments"   "course_item_assessments"         
#>